cuda_prores: fix grey output via power-of-2 slice decomposition
ProRes splits remainder MB columns into power-of-two slices rather than equal-width slices with one clamped remainder. Replicate FFmpeg's halving walk in the demuxer (slices_per_row), entropy kernel, IDCT kernel (new mb_width param + all 6 call sites), and decode_alpha_to_host. Fixes grey/blocky output on non-16-aligned widths (e.g. 2604x1470). Also keeps prior qscale remap, ceil mb_width, uint16 q_scale, and the 4444 alpha heap-overflow crash fix.

+++ src/modules/cuda_prores/cuda/cuda_prores_decode.h
@@ -47,7 +47,7 @@ struct ProResDecodeCtx {
     uint32_t* d_slice_starts;    // [num_slices] byte offset from d_bitstream start
     uint16_t* d_slice_sizes;     // [num_slices] slice byte sizes
     int16_t*  d_dec_coeffs;      // [num_slices × coeff_stride] entropy decode output
-    uint8_t*  d_q_scales;        // [num_slices] q_scale per slice
+    uint16_t* d_q_scales;        // [num_slices] q_scale per slice (FFmpeg-remapped, may exceed 255)
     int16_t*  d_y;               // [height × width]  planar luma
     int16_t*  d_cb;              // [height × width/2] planar Cb  (422), [height × width] (4444)
     int16_t*  d_cr;              // as d_cb
